Component: Move bulk flags update to component
Doing that on project level has serious perfomance impact on big
projects.

Issue #3664
